pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wordflash"
version = "0.1.0"
description = "Generate Anki flashcards from YAML word lists with images and audio"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "WordFlash Contributors" }
]
requires-python = ">=3.8"

dependencies = [
    "pyyaml>=6.0,<7.0",
    "requests>=2.31.0,<3.0.0",
    "gtts>=2.4.0,<3.0.0",
    "genanki>=0.13.0,<1.0.0",
    "pillow>=10.0.0,<11.0.0",
    "tinydb>=4.8.0,<5.0.0",
]

[project.scripts]
wordflash = "wordflash.main:main"
wordflash-analyze = "wordflash.analyze_vocabulary:main"

[tool.hatch.build.targets.wheel]
packages = ["src/wordflash"]
